Delete unused variable and try to fix error in sigterm handling

+++ server/accepter/main.py
@@ -3,6 +3,7 @@ import multiprocessing as mp
 import json
 from MOM import MOM
 import signal
+import errno
 import logging
 
 cluster_type = "accepter"
@@ -26,7 +27,8 @@ class Accepter():
         self.has_to_close = False
         self.previous_stage_size = self.middleware.get_previous_stage_size()
 
-        signal.signal(signal.SIGTERM, self.__handle_signal)
+        # signal.signal(signal.SIGTERM, self.__handle_signal)
+        signal.signal(signal.SIGTERM, self.send_close_signal)
 
     def send_general(self, message):
         self.middleware.send_general(message)
@@ -36,33 +38,48 @@ class Accepter():
 
     def process_received_message(self, ch, method, properties, body):
         response = json.loads(body)
-        if response == None:
-            self.received_eofs += 1
-            if self.received_eofs == self.previous_stage_size:
-                send_json(self.socket, { "finished": True })
-                self.has_to_close = True
-        else:
-            sender = response["type"]
-            if sender == "duplication_filter":
-                received_tuple = response["tuple"]
-                send_json(self.socket, { "type": "first_query", "value": received_tuple, "finished": False })
-            elif sender == "thumbnails_downloader":
-                image_data = response["img_data"]
-                send_json(self.socket, { "type": "second_query", "value": image_data, "finished": False })
-            elif sender == "max_views_day":
-                max_day = response["max_day"]
-                send_json(self.socket, { "type": "third_query", "value": max_day, "finished": False })
+
+        try:
+            if response == None:
+                self.received_eofs += 1
+                if self.received_eofs == self.previous_stage_size:
+                    send_json(self.socket, { "finished": True })
+                    self.has_to_close = True
             else:
-                raise ValueError(f"Unexpected sender: {sender}")
+                sender = response["type"]
+                if sender == "duplication_filter":
+                    received_tuple = response["tuple"]
+                    send_json(self.socket, { "type": "first_query", "value": received_tuple, "finished": False })
+                elif sender == "thumbnails_downloader":
+                    image_data = response["img_data"]
+                    send_json(self.socket, { "type": "second_query", "value": image_data, "finished": False })
+                elif sender == "max_views_day":
+                    max_day = response["max_day"]
+                    send_json(self.socket, { "type": "third_query", "value": max_day, "finished": False })
+                else:
+                    raise ValueError(f"Unexpected sender: {sender}")
+        except socket.error as e:
+            if e.errno == errno.EPIPE:
+                print(f"Client has already closed the connection")
+            else:
+                print(f"Caught unexpected exception while receiving message from server: {str(e)}")
+            # finished = True
+            self.send_close_signal()
+
 
         if self.has_to_close:
             self.middleware.close()
             print("Closed MOM")
 
-    def __handle_signal(self, *args): # To prevent double closing 
+    # def __handle_signal(self, *args): # To prevent double closing 
+    #     self.has_to_close = True
+    #     for process in self.child_processes:
+    #         process.terminate()
+    def send_close_signal(self, *args): # To prevent double closing 
+        if not self.has_to_close:
+            for process in self.child_processes:
+                process.terminate()
         self.has_to_close = True
-        for process in self.child_processes:
-            process.terminate()
 
 def read_json(skt: socket):
     return json.loads(__read_string(skt))
